requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
Django==1.9.7
djangorestframework==3.3.3
drf-extensions==0.2.8
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pickleshare==0.7.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
